Reorganized transcription
